+++ nodes/Autotask/resources/aiHelper/index.ts
@@ -1,0 +1,2 @@
+export { aiHelperFields } from './description';
+export { executeAiHelperOperation } from './execute';